mapper/requirements.txt
# Function dependencies
google-cloud-storage
google-cloud-pubsub